+++ backend/main-service/src/main/java/io/blog/devlog/domain/blog/repository/InfoRepository.java
@@ -1,6 +1,6 @@
-package io.blog.devlog.domain.info.repository;
+package io.blog.devlog.domain.blog.repository;
 
-import io.blog.devlog.domain.info.model.Info;
+import io.blog.devlog.domain.blog.model.Info;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
